+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/EventDtos.kt
@@ -143,6 +143,7 @@ internal data class MessageDeletedEventDto(
     val message: DownstreamMessageDto,
     val hard_delete: Boolean?,
     val channel_message_count: Int? = null,
+    val deleted_for_me: Boolean? = null,
 ) : ChatEventDto()
 
 @JsonClass(generateAdapter = true)