+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/api2/mapping/EventMappingTestArguments.kt
@@ -210,6 +210,7 @@ internal object EventMappingTestArguments {
     private val REMINDER = Mother.randomDownstreamReminderDto()
     private val AI_MESSAGE_ID = randomString()
     private val AI_STATE = randomString()
+    private val DELETED_FOR_ME = randomBoolean()
 
     // BEGIN: DTO Models
 
@@ -416,6 +417,7 @@ internal object EventMappingTestArguments {
         user = USER,
         message = MESSAGE,
         hard_delete = HARD_DELETE,
+        deleted_for_me = DELETED_FOR_ME,
     )
 
     private val messageReadDto = MessageReadEventDto(
@@ -1042,6 +1044,7 @@ internal object EventMappingTestArguments {
         message = with(domainMapping) { messageDeletedDto.message.toDomain() },
         hardDelete = messageDeletedDto.hard_delete ?: false,
         channelMessageCount = messageDeletedDto.channel_message_count,
+        deletedForMe = messageDeletedDto.deleted_for_me ?: false,
     )
 
     private val messageRead = MessageReadEvent(