+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add support for deleting messages only for the current user. [#5967](https://github.com/GetStream/stream-chat-android/pull/5967)
 
 ### ⚠️ Changed
 
@@ -27,6 +28,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add support for deleting messages only for the current user. [#5967](https://github.com/GetStream/stream-chat-android/pull/5967)
 
 ### ⚠️ Changed
 
@@ -34,6 +36,7 @@
 
 ## stream-chat-android-state
 ### 🐞 Fixed
+- Fix poll state getting overridden by `message.new` events. [#5963](https://github.com/GetStream/stream-chat-android/pull/5963)
 - Fix polls not updated live when they are a thread parent message. [#5968](https://github.com/GetStream/stream-chat-android/pull/5968)
 
 ### ⬆️ Improved